Move import_coord_[12]d to python/ctsm/utils.py.
